task a1-4
