+++ src/app/app.module.ts
@@ -4,6 +4,7 @@ import { AppComponent } from './components/parent/app.component';
 import { MetaComponent } from './components/meta/meta.component';
 import { PreviewComponent } from './components/preview/preview.component';
 import { SectionComponent } from './components/section/section.component';
+import { EntryDetailsService } from './components/section/section.service';
 
 @NgModule({
   declarations: [
@@ -15,7 +16,7 @@ import { SectionComponent } from './components/section/section.component';
   imports: [
     BrowserModule
   ],
-  providers: [],
+  providers: [EntryDetailsService],
   bootstrap: [AppComponent]
 })
 export class AppModule { }
