+++ src/app/components/section/section.component.ts
@@ -1,4 +1,6 @@
 import { Component, HostBinding, HostListener, Input } from '@angular/core';
+import { EntryDetailsService } from '../section/section.service';
+import { ISection } from '../section/section.service';
 
 @Component({
   selector: 'sectionSelector',
@@ -7,34 +9,15 @@ import { Component, HostBinding, HostListener, Input } from '@angular/core';
 })
 export class SectionComponent {
 
-  selectedItem: ISection;
+  sections: ISection[];
 
-  sections: ISection[] = [
-    { Id: 0, Label: 'Metadata', IsActive: false },
-    { Id: 1, Label: 'Thumbnails', IsActive: false },
-    { Id: 2, Label: 'AccessControl', IsActive: false },
-    { Id: 3, Label: 'Scheduling', IsActive: false },
-    { Id: 4, Label: 'Flavors > Type', IsActive: false },
-    { Id: 5, Label: 'Captions', IsActive: false },
-    { Id: 6, Label: 'Live', IsActive: false },
-    { Id: 7, Label: 'Related', IsActive: false }
-  ];
+  constructor(private sectionItemService: EntryDetailsService) {
+    this.sections = this.sectionItemService.get();
+  }
 
   changeSelection(section: ISection) {
-    // set rest of items to inactive
-    this.sections.forEach(x => x.IsActive = false);
-    this.selectedItem = section;
-
-    // set item to active
-    section.IsActive = true;
-    console.log(this.sections)
+    this.sectionItemService.activateSection(section);
   }
-}
-
-export interface ISection {
-  IsActive: boolean;
-  Id: number;
-  Label: string;
 }
